api/v1alpha2: Refactor validation into separate files
Refactors the validation package into separate files to make it
easier to manage as the functionality grows.
